requirements.txt
cx_Oracle==7.1.2
fhirstore==0.3.4
Flask-Cors==3.0.8
jsonschema==3.0.2
pandas==0.25.3
psycopg2-binary==2.8.3
pymongo==3.9.0
PyYAML==5.1.2
requests==2.21.0
SQLAlchemy==1.3.13

